+++ src/PIL/BmpImagePlugin.py
@@ -52,6 +52,10 @@ def _accept(prefix):
     return prefix[:2] == b"BM"
 
 
+def _dib_accept(prefix):
+    return i32(prefix[:4]) in [12, 40, 64, 108, 124]
+
+
 # =============================================================================
 # Image plugin for the Windows BMP format.
 # =============================================================================
@@ -101,53 +105,52 @@ class BmpImageFile(ImageFile.ImageFile):
         # --------------------------------------------- Windows Bitmap v2 to v5
         # v3, OS/2 v2, v4, v5
         elif file_info['header_size'] in (40, 64, 108, 124):
-            if file_info['header_size'] >= 40:  # v3 and OS/2
-                file_info['y_flip'] = i8(header_data[7]) == 0xff
-                file_info['direction'] = 1 if file_info['y_flip'] else -1
-                file_info['width'] = i32(header_data[0:4])
-                file_info['height'] = (i32(header_data[4:8])
-                                       if not file_info['y_flip']
-                                       else 2**32 - i32(header_data[4:8]))
-                file_info['planes'] = i16(header_data[8:10])
-                file_info['bits'] = i16(header_data[10:12])
-                file_info['compression'] = i32(header_data[12:16])
-                # byte size of pixel data
-                file_info['data_size'] = i32(header_data[16:20])
-                file_info['pixels_per_meter'] = (i32(header_data[20:24]),
-                                                 i32(header_data[24:28]))
-                file_info['colors'] = i32(header_data[28:32])
-                file_info['palette_padding'] = 4
-                self.info["dpi"] = tuple(
-                    map(lambda x: int(math.ceil(x / 39.3701)),
-                        file_info['pixels_per_meter']))
-                if file_info['compression'] == self.BITFIELDS:
-                    if len(header_data) >= 52:
-                        for idx, mask in enumerate(['r_mask',
-                                                    'g_mask',
-                                                    'b_mask',
-                                                    'a_mask']):
-                            file_info[mask] = i32(
-                                header_data[36 + idx * 4:40 + idx * 4]
-                            )
-                    else:
-                        # 40 byte headers only have the three components in the
-                        # bitfields masks, ref:
-                        # https://msdn.microsoft.com/en-us/library/windows/desktop/dd183376(v=vs.85).aspx
-                        # See also
-                        # https://github.com/python-pillow/Pillow/issues/1293
-                        # There is a 4th component in the RGBQuad, in the alpha
-                        # location, but it is listed as a reserved component,
-                        # and it is not generally an alpha channel
-                        file_info['a_mask'] = 0x0
-                        for mask in ['r_mask', 'g_mask', 'b_mask']:
-                            file_info[mask] = i32(read(4))
-                    file_info['rgb_mask'] = (file_info['r_mask'],
-                                             file_info['g_mask'],
-                                             file_info['b_mask'])
-                    file_info['rgba_mask'] = (file_info['r_mask'],
-                                              file_info['g_mask'],
-                                              file_info['b_mask'],
-                                              file_info['a_mask'])
+            file_info['y_flip'] = i8(header_data[7]) == 0xff
+            file_info['direction'] = 1 if file_info['y_flip'] else -1
+            file_info['width'] = i32(header_data[0:4])
+            file_info['height'] = (i32(header_data[4:8])
+                                   if not file_info['y_flip']
+                                   else 2**32 - i32(header_data[4:8]))
+            file_info['planes'] = i16(header_data[8:10])
+            file_info['bits'] = i16(header_data[10:12])
+            file_info['compression'] = i32(header_data[12:16])
+            # byte size of pixel data
+            file_info['data_size'] = i32(header_data[16:20])
+            file_info['pixels_per_meter'] = (i32(header_data[20:24]),
+                                             i32(header_data[24:28]))
+            file_info['colors'] = i32(header_data[28:32])
+            file_info['palette_padding'] = 4
+            self.info["dpi"] = tuple(
+                map(lambda x: int(math.ceil(x / 39.3701)),
+                    file_info['pixels_per_meter']))
+            if file_info['compression'] == self.BITFIELDS:
+                if len(header_data) >= 52:
+                    for idx, mask in enumerate(['r_mask',
+                                                'g_mask',
+                                                'b_mask',
+                                                'a_mask']):
+                        file_info[mask] = i32(
+                            header_data[36 + idx * 4:40 + idx * 4]
+                        )
+                else:
+                    # 40 byte headers only have the three components in the
+                    # bitfields masks, ref:
+                    # https://msdn.microsoft.com/en-us/library/windows/desktop/dd183376(v=vs.85).aspx
+                    # See also
+                    # https://github.com/python-pillow/Pillow/issues/1293
+                    # There is a 4th component in the RGBQuad, in the alpha
+                    # location, but it is listed as a reserved component,
+                    # and it is not generally an alpha channel
+                    file_info['a_mask'] = 0x0
+                    for mask in ['r_mask', 'g_mask', 'b_mask']:
+                        file_info[mask] = i32(read(4))
+                file_info['rgb_mask'] = (file_info['r_mask'],
+                                         file_info['g_mask'],
+                                         file_info['b_mask'])
+                file_info['rgba_mask'] = (file_info['r_mask'],
+                                          file_info['g_mask'],
+                                          file_info['b_mask'],
+                                          file_info['a_mask'])
         else:
             raise IOError("Unsupported BMP header type (%d)" %
                           file_info['header_size'])
@@ -176,6 +179,7 @@ class BmpImageFile(ImageFile.ImageFile):
             SUPPORTED = {
                 32: [(0xff0000, 0xff00, 0xff, 0x0),
                      (0xff0000, 0xff00, 0xff, 0xff000000),
+                     (0xff, 0xff00, 0xff0000, 0xff000000),
                      (0x0, 0x0, 0x0, 0x0),
                      (0xff000000, 0xff0000, 0xff00, 0x0)],
                 24: [(0xff0000, 0xff00, 0xff)],
@@ -184,6 +188,7 @@ class BmpImageFile(ImageFile.ImageFile):
             MASK_MODES = {
                 (32, (0xff0000, 0xff00, 0xff, 0x0)): "BGRX",
                 (32, (0xff000000, 0xff0000, 0xff00, 0x0)): "XBGR",
+                (32, (0xff, 0xff00, 0xff0000, 0xff000000)): "RGBA",
                 (32, (0xff0000, 0xff00, 0xff, 0xff000000)): "BGRA",
                 (32, (0x0, 0x0, 0x0, 0x0)): "BGRA",
                 (24, (0xff0000, 0xff00, 0xff)): "BGR",
@@ -196,7 +201,7 @@ class BmpImageFile(ImageFile.ImageFile):
                     raw_mode = MASK_MODES[
                         (file_info["bits"], file_info["rgba_mask"])
                     ]
-                    self.mode = "RGBA" if raw_mode in ("BGRA",) else self.mode
+                    self.mode = "RGBA" if "A" in raw_mode else self.mode
                 elif (file_info['bits'] in (24, 16) and
                       file_info['rgb_mask'] in SUPPORTED[file_info['bits']]):
                     raw_mode = MASK_MODES[
@@ -291,7 +296,11 @@ SAVE = {
 }
 
 
-def _save(im, fp, filename):
+def _dib_save(im, fp, filename):
+    _save(im, fp, filename, False)
+
+
+def _save(im, fp, filename, bitmap_header=True):
     try:
         rawmode, bits, colors = SAVE[im.mode]
     except KeyError:
@@ -306,14 +315,15 @@ def _save(im, fp, filename):
 
     stride = ((im.size[0]*bits+7)//8+3) & (~3)
     header = 40  # or 64 for OS/2 version 2
-    offset = 14 + header + colors * 4
     image = stride * im.size[1]
 
     # bitmap header
-    fp.write(b"BM" +                      # file type (magic)
-             o32(offset+image) +          # file size
-             o32(0) +                     # reserved
-             o32(offset))                 # image data offset
+    if bitmap_header:
+        offset = 14 + header + colors * 4
+        fp.write(b"BM" +                      # file type (magic)
+                 o32(offset+image) +          # file size
+                 o32(0) +                     # reserved
+                 o32(offset))                 # image data offset
 
     # bitmap info header
     fp.write(o32(header) +                # info header size
@@ -352,3 +362,10 @@ Image.register_save(BmpImageFile.format, _save)
 Image.register_extension(BmpImageFile.format, ".bmp")
 
 Image.register_mime(BmpImageFile.format, "image/bmp")
+
+Image.register_open(DibImageFile.format, DibImageFile, _dib_accept)
+Image.register_save(DibImageFile.format, _dib_save)
+
+Image.register_extension(DibImageFile.format, ".dib")
+
+Image.register_mime(DibImageFile.format, "image/bmp")